Sync Adapter well implemented, bugs for today's layout and list fixed
