+++ app/src/main/AndroidManifest.xml
@@ -37,7 +37,6 @@
             android:name=".data.WeatherProvider"
             android:exported="false"
             android:syncable="true"/>
-        <service android:name=".services.SunshineService"/>
         <receiver android:name=".services.SunshineService$AlarmReceiver"/>
         <service android:name=".services.AuthenticatorService">
             <intent-filter>
