+++ app/src/main/res/layout/list_item_forecast.xml
@@ -1,7 +1,7 @@
 <LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:minHeight="?android:attr/listPreferredItemHeight"
-    android:padding="16dp"
+    android:padding="@dimen/activity_vertical_margin"
     android:background="@drawable/touch_selector"
     android:layout_width="match_parent"
     android:layout_height="wrap_content">
@@ -11,6 +11,9 @@
         android:layout_height="wrap_content"
         android:id="@+id/imageIconWeather"
         android:scaleType="centerInside"
+        android:layout_gravity="center"
+        android:layout_marginRight="@dimen/space"
+        android:layout_marginEnd="@dimen/space"
         android:src="@drawable/ic_clear" />
 
     <LinearLayout
