+++ app/src/main/res/layout/list_item_forecast_today.xml
@@ -3,7 +3,8 @@
     android:orientation="horizontal"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:padding="16dp"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
     android:minHeight="?android:attr/listPreferredItemHeight"
     android:gravity="center_vertical">
