+++ app/src/main/res/values/dimens.xml
@@ -2,5 +2,7 @@
     <!-- Default screen margins, per the Android Design guidelines. -->
     <dimen name="activity_horizontal_margin">16dp</dimen>
     <dimen name="activity_vertical_margin">16dp</dimen>
+    <dimen name="double_vertical_padding">32dp</dimen>
     <dimen name="normal_padding">10dp</dimen>
+    <dimen name="space">6dp</dimen>
 </resources>
